First web version
First functional version of the website.
Include a home page + a page to display individual game.

At the moment, the games are manualy include in the public folder,
the game name is hardcoded. It's not a big deal at the moment, but
I want to change it to something more modular.

Signed-off-by: Maxime Millette-Coulombe <mmillettecoulombe@gmail.com>